refactor(auth): update request types to use DTOs and rename handler imports
