+++ internal/http/router/auth.go
@@ -1,12 +1,12 @@
 package router
 
 import (
-	"key-haven-back/internal/handler"
+	handler2 "key-haven-back/internal/http/handler"
 
 	"github.com/gofiber/fiber/v3"
 )
 
-func RegisterRoutes(app *fiber.App, authHandler *handler.AuthHandler, passwordHandler *handler.PasswordHandler) {
+func RegisterRoutes(app *fiber.App, authHandler *handler2.AuthHandler, passwordHandler *handler2.PasswordHandler) {
 	auth := app.Group("/auth")
 	auth.Post("/register", authHandler.Register)
 	auth.Post("/login", authHandler.Login)
