+++ internal/http/router/router.go
@@ -1,12 +1,12 @@
 package router
 
 import (
-	"key-haven-back/internal/handler"
+	handler2 "key-haven-back/internal/http/handler"
 
 	"github.com/gofiber/fiber/v3"
 )
 
-type RegisterRoutesFunc func(app *fiber.App, authHandler *handler.AuthHandler, passwordHandler *handler.PasswordHandler)
+type RegisterRoutesFunc func(app *fiber.App, authHandler *handler2.AuthHandler, passwordHandler *handler2.PasswordHandler)
 type RegisterSwaggerRoutesFunc func(app *fiber.App)
 
 func RegisterRoutesFuncProvider() RegisterRoutesFunc {
